refactor(tts): migrate OpenAI TTS to shared TextToSpeechModel interface and standardize on Double for speed
  BREAKING CHANGE: OpenAI TTS now uses shared interfaces instead of provider-specific classes

  - Replace deprecated OpenAI-specific classes (SpeechModel, SpeechPrompt, SpeechResponse,
    StreamingSpeechModel) with shared interfaces (TextToSpeechModel, TextToSpeechPrompt,
    TextToSpeechResponse, StreamingTextToSpeechModel)
  - Update interface hierarchy: TextToSpeechModel now extends StreamingTextToSpeechModel
    for consistency with ChatModel pattern
  - Add @FunctionalInterface annotation to StreamingTextToSpeechModel
  - Migrate speed property from Float to Double across all OpenAI TTS components for
    consistency with ElevenLabs API and better type uniformity
  - Update OpenAiAudioSpeechOptions to implement TextToSpeechOptions interface
  - Add TextToSpeechModelTests with 7 comprehensive unit tests for interface behavior
  - Improve null safety in TextToSpeechModel.call(String) default method
  - Update documentation to reflect new interface hierarchy and Double usage
  - Add migration guide to documentation for upgrading from deprecated classes

+++ models/spring-ai-openai/src/main/java/org/springframework/ai/openai/api/OpenAiAudioApi.java
@@ -396,7 +396,7 @@ public class OpenAiAudioApi {
 		@JsonProperty("input") String input,
 		@JsonProperty("voice") String voice,
 		@JsonProperty("response_format") AudioResponseFormat responseFormat,
-		@JsonProperty("speed") Float speed) {
+		@JsonProperty("speed") Double speed) {
 		// @formatter:on
 
 		public static Builder builder() {
@@ -491,7 +491,7 @@ public class OpenAiAudioApi {
 
 			private AudioResponseFormat responseFormat = AudioResponseFormat.MP3;
 
-			private Float speed;
+			private Double speed;
 
 			public Builder model(String model) {
 				this.model = model;
@@ -518,7 +518,7 @@ public class OpenAiAudioApi {
 				return this;
 			}
 
-			public Builder speed(Float speed) {
+			public Builder speed(Double speed) {
 				this.speed = speed;
 				return this;
 			}
